Move lib into src/
