+++ packages/server-web/src/lib/label-extractor.js
