+++ packages/server-web/src/lib/label-extractor.spec.js
