+++ packages/server-web/src/lib/string-to-color.js
