+++ packages/server-web/src/lib/string-to-color.spec.js
